feat(stages): implement StageRepository for updating stage summaries and progress tracking
