+++ package.json
@@ -57,7 +57,8 @@
     "prom-client": "^15.1.3",
     "reflect-metadata": "^0.2.2",
     "tsyringe": "^4.8.0",
-    "xstate": "^5.19.2"
+    "xstate": "^5.19.2",
+    "lodash": "^4.17.21"
   },
   "devDependencies": {
     "@commitlint/cli": "^19.8.0",
@@ -89,7 +90,6 @@
     "jest-extended": "^4.0.2",
     "jest-html-reporters": "^3.1.7",
     "jest-openapi": "^0.14.2",
-    "lodash": "^4.17.21",
     "prettier": "^3.3.3",
     "pretty-quick": "^4.1.1",
     "prisma": "^6.6.0",
