+++ src/db/createConnection.ts
@@ -35,7 +35,6 @@ export const createConnectionOptions = (dbConfig: commonDbFullV1Type): PoolConfi
 
 export function createPrismaClient(poolConfig: PoolConfig, schema: string): PrismaClient {
   const adapter = new PrismaPg(poolConfig, { schema });
-  // const prisma = new PrismaClient({ adapter, log:[{ emit:'stdout', level:'query'}] });
   const prisma = new PrismaClient({ adapter });
 
   return prisma;
